requirements.txt
flask
pdfminer.six
nltk
spacy
langchain
langchain-chroma
langchain-huggingface
sentence-transformers
requests
chromadb
